Extend the logic for linux
The machine id is looked in "canonical" locations. If not value is found, one is generated and persisted.
The machine id is looked in:
  - the file pointed by the `MACHINE_ID_FILE` env var
  - `/var/lib/dbus/machine-id`
  - `/etc/machine-id`
  - `$HOME/.config/machine-id`

If no such file is found, a random uuid is generated and persisted in the first
writable file among `$MACHINE_ID_FILE`, `/var/lib/dbus/machine-id`, `/etc/machine-id`, `$HOME/.config/machine-id`.

If there is an error reading _all_ the files an empty string is returned.

The logic implemented is a variation of the one described in https://github.com/denisbrodbeck/machineid/issues/5#issuecomment-523803164